Stage 1: add DiscreteActionPolicy module
New module src/model/discrete_policy.py integrates:
- Per-cell logit head MLP (neuromod) → K-way categorical
- Shared codebook [K, D] embedding table
- Shared decoder MLP [D → action_dim]

Replaces the old encoder→VQ→decoder pipeline with a single unified
discrete action policy. Supports Gumbel-softmax for phase 1 backprop
and hard Categorical for phase 2 RL. Includes usage tracking and
dead-code reset for codebook quality during bootstrap.

Not yet wired into memory.py — standalone with unit tests only.
12 unit tests all passing.

+++ src/model/memory.py
@@ -1306,7 +1306,10 @@ class MemoryGraph(nn.Module):
                     #     log π = -½ Σ ε_new² - ½·D·log(2πσ²)
                     # where μ_now is the modulator re-forward on mod_inputs
                     # (with current weights).
-                    codes = sampled_flat.detach()
+                    # Store sample in bf16 — action_dim is O(1000) so f32
+                    # storage at high n_calls causes OOM. bf16 has plenty
+                    # of dynamic range for log-prob reconstruction.
+                    codes = sampled_flat.detach().to(torch.bfloat16)
                     quantized = sampled_flat.reshape(BS, NC, -1).to(dt)
                 else:
                     # Normalize, encode, sample codes, decode, un-normalize (all f32)
